refactor: move type to types directory
