+++ src/index.ts
@@ -3,4 +3,4 @@ export { dividerFirst } from '@/core/divider-first';
 export { dividerLast } from '@/core/divider-last';
 export { dividerLoop } from '@/core/divider-loop';
 export { dividerNumberString } from '@/core/divider-number-string';
-export type { DividerResult } from '@/core/types';
+export type { DividerResult } from '@/types';
